WebApp/requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.30.1
paho-mqtt==1.6.1
python-telegram-bot==20.7
requests==2.32.3
